Safari-611.1.10/LayoutTests/media/audio-constructor-expected.txt

Test that Audio() object loads the resource after src attribute is set and load() is called.

EXPECTED (audio instanceof HTMLAudioElement == 'true') OK
RUN(audio.load())
EVENT(loadstart)
EXPECTED (relativeURL(audio.currentSrc) == 'content/test.wav') OK
EVENT(canplaythrough)
END OF TEST
